Format call stack as in Python
Summary:
# What is different

* two lines per frame
* only keep start line number, drop start column, end line and column
* include snippet

## Remaining issues

* on error, line snippet with failure is printed twice (once in the stack, and another as part of error). The issue existed before, but there was no snippet, only file and line numbers
* function names are often too long
  * when printing partial in call stack, we should drop the arguments
  * when printing starlark functions, we should include only local name, not full id, i. e. `_my_library`, not `//tools/build_defs:my_library.bzl._my_library`
* bottom frame is called `<module>`, could be something more meaningful
* in error messages we use paths, not module ids (but this is not starlark issue)

## Top function name is not printed

Possibly significant difference between old and new output is that top function name is no longer printed (but the snippet which calls it is printed). For example, when the top function is `cxx_library`, before, function name is printed:

```
  cxx_library() (called from //tools/build_defs:fb_native_wrapper.bzl:258:5-33)
```

and now it is not:

```
    File //tools/build_defs:fb_native_wrapper.bzl, line 258, in //tools/build_defs:fb_native_wrapper.bzl._fb_native_cxx_library
      native.cxx_library(**kwargs)
```

Here if `cxx_library` was called using an alias (`foo = native.cxx_library`), it wouldn't be possible to find it's name.

There's no way to put it Python-compatible output. Python doesn't print top function name either. For example, for program:

```
def foo():
    test = int
    test('a')

foo()
```

Python prints:

```
Traceback (most recent call last):
  File "./a.py", line 9, in <module>
    foo()
  File "./a.py", line 7, in foo
    test('a')
ValueError: invalid literal for int() with base 10: 'a'
```

We can only learn that top function name is `int` from the error message produced by the function itself.

Reviewed By: ndmitchell

Differential Revision: D35806696

fbshipit-source-id: 673823710da5038d2e1f009dc76a11ebe90d02ae