Link to the lint format
Summary:
Hard to find what is this without explanation.

This links to FB internal page, but that's fine, because the format is internal too.

However, we should get rid of this linter (output formatting, not the implementation) from starlark project, and make it either
* standalone binary outside of starlark, or better
* part of buck

Making linter part of buck I think would be the most appropriate because the linter would be able to also load dependencies and lint better.

Reviewed By: ndmitchell

Differential Revision: D36232461

fbshipit-source-id: 0d2e5a1df00ef5ec6cd11e152f34ee1951ccb065