Update some tests to be more readable
Summary:
We have several tests for LSP where we will want to refer to ranges / ResolvedSpans within a file. This is tedious and fragile if we hardcode line/columns, so add in a simple format to let us embed identifiers in the sample source and retrieve those spans by name.

The format is simple, just add `<ID>` / `</ID>` around the text that you want the range for, and then resolve it with `FixtureWithRanges::span`.

Reviewed By: ndmitchell

Differential Revision: D36255954

fbshipit-source-id: 7ccdc8c2a5ee170796cea8cf97668ad8b6a5447e