Add simple AstModule cache to lsp
Summary:
Add a simple cache of the ASTModule for files when we parse them. This is needed so we can jump to definitions before users have saved their files (we update this cache on change/open of the file).

As all of the methods in eval already parse the ASTModule out, change those methods to return both the Message iterator, and the ast, so that we're not parsing twice.

Reviewed By: ndmitchell

Differential Revision: D36169668

fbshipit-source-id: 1511a35d3679bda3a49c12cb8092cfeb2158beff