Goto local definitions of variables.
Summary:
Add the logic to find the definition of a symbol underneath a cursor. This runs through the analysis::bind logic, so it follows scopes properly, too.

Right now builtins / globals are not handled, object properties are not handled (e.g. clicking on `bar` in `foo.bar` will not jump anywhere), and symbols that are loaded only jump to the symbol parameter of the correct "load" statement.

Reviewed By: ndmitchell

Differential Revision: D36169667

fbshipit-source-id: 85e8727e74f480c3dbfe81c40ca88d88f1103783