Add extra information to bind::Assigner for loads
Summary: If a symbol is loaded, pass both the name in that file, and the name of the file, into the Assigner object. This is being done so that if a user tries to look up a definition, and it was imported from a load statement, we can then attempt to resolve the original file and symbol, rather than just jumping to the load() line.

Reviewed By: ndmitchell, stepancheg

Differential Revision: D36295062

fbshipit-source-id: 67f48a17033ec3c7a3bee7d22a078784ea590778