Create LspContext / LspEvalResult
Summary: Create an LSP specific context that will handle things like loading files, resolving load statements, etc. This also removes the dependency in the LSP on the context in bin/eval.rs, which can just implement the LSPContext trait. This is being done to move lsp code into the library.

Reviewed By: stepancheg

Differential Revision: D36292918

fbshipit-source-id: 68e1395ad887b0c3028157945de6733e858f024b