Move eval tests
Summary:
Move `eval::tests` to just `tests`.

Some of the tests are not evaluation tests (derive tests for example), and for some tests it is not clear what they tests exactly: parse, evaluation, runtime or full stack?

So just move all the tests.

Generally the idea is this. If a test tests something specific, like substring behaviour, the test should next to substring implementation. If one test affects several submodules, place it into top level `tests` module.

The motivation for this stack is to make it easier to find specific tests. Hope updated structure makes it easier.

Reviewed By: ndmitchell

Differential Revision: D36345849

fbshipit-source-id: 083b85dfe8fef7b3b2405255a2b44a62b037f979