In derive, function parameter default value should be Expr, not Pat
Summary:
This is more correct: if pattern is not an expression, we fail at macro run time instead of users of macro compilation.

For example,

```
(x @ y, z @ w)
```

is a valid pattern, but this is not a valid expression.

This is also needed for the following diff D36393574, but this is useful as is.

Reviewed By: ndmitchell

Differential Revision: D36393573

fbshipit-source-id: 45c09b0af376ebc83ad9352a00d75aaa465a144e