Invoke frozen def with resolved argument names
Summary:
Consider this code:

```
def foo(a, **kwargs): ...

foo(a=1, b=2)
```

Before this diff `foo` was called with two symbols — for `"a"` and `"b"` which were resolved with hashmap lookups in `DefGen`.

Now `"a"` and `"b"` are resolved to indices at compilation time (`Some(0)` and `None`).

# Previous diffs in the stack

are mostly mechanical refactorings to enable this.

Reviewed By: ndmitchell

Differential Revision: D33000505

fbshipit-source-id: 2c200c43df2660ffaf86f548b1280ad918a376c4