Look up an exported symbol by name
Summary: Add a quick function to look up the span for exported symbols. This is used in future diffs to look up globals in files that are loaded into a file that a user is looking for a definition in, so that we can quickly get to the original definition.

Reviewed By: ndmitchell, stepancheg

Differential Revision: D36425721

fbshipit-source-id: 06d7daace59c6dece9c366374f379043849f0466