Make LSP context resolve paths for load statements, and return file contents
Summary:
Add some extra methods to the LSP context to allow resolving load paths and the contents of those paths.

URL is used as the type for resolving load paths so that the LSP can do any caching of intermediate file state that it needs to more easily. (e.g. right now it stores file contents sent from the client in a hash map based on the URI)

Reviewed By: ndmitchell, stepancheg

Differential Revision: D36425724

fbshipit-source-id: cc35f08e81b754cacad00f4b78d5ea5c46073dbb