Accept &Heap and &mut Evaluator params in native functions
Summary:
Native functions now accept `heap: &Heap` or `eval: &mut Evaluator` parameters.

Following stack up to D36749507 makes these parameters mandatory if `heap` or `eval` variable are used.

There are two goals for this stack:
* trade code read speed to write speed: it is required to type more to write functions, but functions are easier to read: it's clear what `eval` or `heap` variables are, or it is possible to Cmd-Click on `Evaluator` to get to the type definition
* there's work I'm trying to do which depends on it

### Long story

First, we should replace `&mut Evaluator` with `CallContext`. Simple as-is implementation results in borrow checker errors, which should be easier to deal with if function explicitly requests what it needs. In particular, before this diff we unconditionally create both `heap` and `eval` local variables, and how it works, it is mystery to me, and with implementation of `CallContext` it is borrow checker error. With this stack, a function must explicitly request heap or evaluator, and it is not permitted to request both.

`CallContext` is needed restrict what `Evaluator` exposes. In particular, `Evaluator` should not expose heap: heap should be available only during evaluation — via `CallContext`.

After that, `Evaluator` should not grab `Module` during construction, but should accept module in `eval_module` function. So lifetime of `Evaluator` will not be tied to lifetime of module.

And when `Evaluator` lifetime is not tied to `Module` lifetime, we can pass `Evaluator` to module freeze function.

Which is needed to speculatively execute functions during freeze. Currently we speculatively execute functions only during module evaluation.

I have prototyped inlining of functions with constant arguments. It resulted in no performance gain. I suspect that speculative evaluation on freeze will make inlining of functions with const arguments useful.

Reviewed By: ndmitchell

Differential Revision: D36749508

fbshipit-source-id: fb3630d68d7c132aa45d40f6f8415b04363d6f58