BcSlotOut type
Summary:
This is similar to D36865227 (https://github.com/facebookexperimental/starlark-rust/commit/f27554891912e8f3a63557612dfb76fddc38132c), but for output slots.

This is not as helpful for safety, although may prevent some errors.

But makes it more clear what is used for writing and what is for reading, especially helpful in instruction arguments with several slot parameters.

Reviewed By: ndmitchell

Differential Revision: D36891262

fbshipit-source-id: cecbe00514abb397b01ffdadd825059a58b30eeb