Remove special handling of methods in AST compiler
Summary:
Move special handling of methods to optimizer part of the compiler.

This diff changes how `"aaa{}bbb".format(x)` optimization is implemented.

Previously it hooked at getattr, but now it no longer works because when `CallCompiled::call` is called, `"aaa{}bbb".format` (without args) is folded into `BoundMethod`.

Now unnecessary `BoundMethod` is allocated during compilation, but this is not a big deal.

(But as a side effect, `"aaa{}bbb".format(x)` is also optimized on freeze, which is also not a big deal.)

Reviewed By: ndmitchell

Differential Revision: D37260502

fbshipit-source-id: dfcef6ec1f0b399e2acea281b5044793481c406b