Add a method to get the repr of a type that is used in type annotations
Summary:
Add a function that can be used to get a string representation that a user would need to use to indicate the specified type in type annotations in starlark. (e.g. `str.type`, `"artifact"`, `None` on `NoneType`, etc)

This defaults to the value provided to `starlark_type!`, but can be overridden with a new second argument to that macro.

This is mostly used for documentation purposes in future revisions.

Reviewed By: ndmitchell

Differential Revision: D37505625

fbshipit-source-id: 91295e342b0bd08bdef73211f73a59c58f20e85e