Handle failure to read files better
Summary:
Change StringLiteralResult to make the lookup function optional, and change the handling when it *is* a starlark file so that any errors are ignored, and we at least jump to the correct file.

This often manifested before as problems when trying to jump to a file that didn't exist, or that wasn't a string.

Reviewed By: ndmitchell

Differential Revision: D37631860

fbshipit-source-id: 307e85732d5150fed0e92cd250e6889970181be6