RawPointer
Summary:
`Pointer<'a>` is logically equivalent to `&'a AValueHeader`.

And `RawPointer` is logically equivalent to `*mut AValueHeader`.

Use `RawPointer` to instead of `usize`, because when using `usize` it is not clear which `usize` is a pointer and which is raw pointer. For example, [here](https://fburl.com/code/e3gqrhnu) one `usize` is index and another `usize` is pointer.

Reviewed By: krallin

Differential Revision: D37606430

fbshipit-source-id: eea2b7aa612dc4a44389e9ed987cbffdc8686fcc