Don't make npm install be a fork bomb
Summary: On my copy of node, running `cd client && npm install` causes it to run npm in the level above (not really sure why - perhaps npm decided it was part of a project somehow?). That then does an install again, which does an install again, and forms an infinite loop. Just tell uses to do the install themselves in `client` to avoid this possibility.

Reviewed By: krallin

Differential Revision: D39083671

fbshipit-source-id: 4a4efc2c288397d1a93a6b6800bf28d4de40e459