Handle getting the definition for a member access
Summary:
If a user wants to get the location of the variable definition when the position they request is part of a dotted member lookup, allow this.

This is done by adding a `DottedDefinition` type, and adding it to the `Definition` enum.

The `root_definition_location` is where the definition of the left-most component of the access is. Right now, this is the only part that is taken into account. In a future diff, the right hand side will be passed through to do further resolution.

Reviewed By: stepancheg

Differential Revision: D38611153

fbshipit-source-id: c06362706ad3ec54882b76ac6cdc58efc6e265fd