Jump to struct member
Summary:
When attempting to find the original definition of a member access, look for a struct with the name of the root element. This must be of the form `Identifier = struct(...)` in the file.

If that's found, attempt to find a named arg with the same name as the clicked member. If the symbol assigned in the arg exists, jump to that.

Reviewed By: stepancheg

Differential Revision: D38611157

fbshipit-source-id: 213375c43ffe6e148b97d3cb8d77290d3577480d