FlameGraphData
Summary:
Intermediate representation of flamegraph.pl data as a tree before writing it to text file.

This diff converts time flame to this intermediate representation and following diff D40281829 does the same for heap flame.

This diff does nothing useful only creates serialization overhead, but it is needed to implement these two features:
* trim too large flame graphs which cannot be opened in browser (e. g. collapse deep nodes which account for less than 0.0001 of total)
* augment flame profiles with non-starlark data in buck2 (e.g. overhead of `ConfiguredTargetNode` data can be included in flame data too)

Reviewed By: IanChilds

Differential Revision: D40281613

fbshipit-source-id: 623469e582aea1a9a6801e93ad4152559f659678