Reimplement SmallMap (again)
Summary:
Change `VecMap` (which is linear part of `SmallMap`) from

```
Vec<(Hash, K, V)>
```

to

```
Vec2<(K, V), Hash>
```

The idea is this: hash is 32 bit. But because of padding, it occupies 64 bits. So by changing layout from

```
H, K, V, H, K, V, ....
```

to

```
K, V, K, V, ..., H, H, ...
```

we save 4 bytes per entry.

At cost of slower operations with `SmallMap`. However `SmallMap` is meant to be small, not fast, so this is probably acceptable.

Basically we trade 2% memory for 0.5% CPU.

Also, additional savings of memory are possible: because hashes are stored in contiguous memory now, we can use SIMD to scan them. So we can increase the amount of entries with scan linearly before allocating the index without hurting much performance (currently we create the index at 12 entries).

Also, we can now implement storing `RawTable` next to `Vec2` data, to reduce the size of empty `SmallMap` from 4 words to 3.

Reviewed By: ndmitchell

Differential Revision: D40557486

fbshipit-source-id: 44ef3dd8de0b72171fa084a417dbf7376c793bae