Hide FrozenDict
Summary:
There are potential optimizations.

For example, we may want to have two frozen `dict` variants:
* without index
* with index

Or for example, we may want to allocate frozen dict content inline, not in malloc, as we do with frozen lists.

To make such changes possible, dict implementation details need to be hidden from public API.

Reviewed By: ndmitchell

Differential Revision: D42410384

fbshipit-source-id: aa50ce449da2ee0600fff163cb43a82da2c51afd