New crate
Summary:
Move `Dupe` implementation into separate crate.

This crate contains:
* `trait Dupe`
* derives `Dupe`, `Dupe_`, `Clone_`, `Copy_`

Parts of "dupe" still live in gazebo (e.g. `duped` function), this is to be fixed later.

The idea is this: gazebo contains too many unrelated parts, some is good, other may be not universally acceptable. Moreover, code quality varies. So when put together, it is not very useful as is as general purpose utility library. Split may help:
* split good parts into independent libraries, each solving own problem
* and move the rest into `buck2_core` or something
* and then archive `gazebo` crate (but keep `gazebo` repo)

Reviewed By: ndmitchell

Differential Revision: D41057314

fbshipit-source-id: ff7c57390dd4d6f6fb2af029fa112bd7518f6bfb