Mark Dupe::dupe inline
Summary:
Was thinking why another change causes measurable perf regression, and found this.

Non-generic functions not marked `#[inline]` are not inlined cross-crate as far as I understand.

We don't do link time optimization when compiling buck2, right?

Additionally, generated `Clone` implementation with `derive(Clone)` is also marked `#[inline]`, so if they do it, they know why they are doing it, so we should do it too.

Reviewed By: ndmitchell

Differential Revision: D42523922

fbshipit-source-id: a4f6ef5b5b3f9b9892f848e9c5d2b112812bc088