Don't smush items of different types into an array
Summary:
The array starts with a variable, and is followed by attributes. That's a really bad plan that leads to really predictable bugs (one of which I fix in the next diff).

I pushed this change halfway through the IDE, where it used references, but stopped when it clones the data anyway.

Reviewed By: stepancheg

Differential Revision: D42577880

fbshipit-source-id: f5f59fe8aef0e4ce4e338c121ad7dd98fb30939b