Add custom starlark print handler that uses console_message
Summary:
Regular Starlark `print()`/`pprint()` does not actually print anything out to the users. This is because `no-buckd` writes directly to stderr rather than the stderr file that the daemon writes to, but the superconsole will overwrite the stderr pipe.

By using `console_message()` for the Starlark print handler, the server can correctly use the right channels to actually emit the Starlark print statements correctly.

Let's add a print handler that does exactly that.

Reviewed By: krallin

Differential Revision: D43466807

fbshipit-source-id: 8de33b8a3fdc901f6670990a4e984cc97045b0df