Expose breakpoint resolution
Summary:
For the buck debugger, we'll send the same breakpoints to many
different evaluators. To support that efficiently, we want to only resolve them
once.

So, expose the breakpoint resolution as a separate thing and then
set_breakpoints can take those resolved breakpoints. This also means that
we don't need get_ast() as part of the DapAdapterClient and different clients
can do that however they want.

Reviewed By: stepancheg

Differential Revision: D43941408

fbshipit-source-id: 6e7a031cce48b8d10715d236bf2ba9c166fdb988