Feature flags to make lsp and readline support optional
This diff moves support for readline (used by the breakpoint console)
and the LSP server behind a cargo feature.  Readline is enabled in the
lib by default to preserve debugging when embedding, while lsp is only
enabled when building the starlark binary.

The motivation is to support a wasm32-unknown-unknown build of
starlark.