Pass the default on to documentation if you can
Summary: If you define a Rust function for Starlark, that has a default, it would be great for that default to show up in the docs. Previously it didn't and we made up None. Now, if it exists, and we can turn it into a FrozenValue, include it in the documentation.

Reviewed By: stepancheg

Differential Revision: D45225574

fbshipit-source-id: 14caaf0ab5de02950033656b4a8c40802ff6215c