Switch to EvaluationCallbacks for before_stmt
Summary:
This means we no longer need to write completely different bytecode to
support before_stmt. In turn, this means we can enable profiling or debugging
without needing to invalidate the entire graph.

I tried many different approaches for this to find one that didn't regress
performance in the disabled case and this is the best result.

For example, I tried:
1. always produce BeforeStmt bytecode and check at
runtime if we can skip it. This was a significant (like >5%) regression
2.  always produce BeforeStmt bytecode, add EvaluationCallbacks so the disabled
case could be optimized by the compiler. This was better, but still a
regression (iirc, something like 3-4%).
3. write two opcodes, select which to use based on EvaluationCallbacks. This was barely better than 2.
4. use the approach in this diff, but with the StmtId in the BcInstrHeader
(rather than in the repr after the args). This was really close (like ~1%
regression).
5.  the approach in this diff. This is no regression.

Checked difference in statement profiling. There are two differences:

1. an if statement with an inlined constant False previously would be hit by
before_stmt and now isn't (because the compiler doesn't produce any
instructions for such a statement)
2. top-level gc instructions now get a before_stmt call, with a span for the
statement that they are preceding

(1) seems like fine behavior. we could feasibly change that case to write a
no-op statement, but it doesn't seem important.
(2) I'd argue this is actually fixing a bug in the current code. I'd hit this
in the past when profiling and it shows up as like a random return statement
potentially deep in some stack would be assigned the time that is spent in gc.
A further improvement (imo) would be to change the gc span to have begin==end
(i.e. point to a zero-length span at the beginning of the statement that
follows the gc) and then we could distinguish it in profiles from the following
statement.

Reviewed By: stepancheg

Differential Revision: D45200183

fbshipit-source-id: b9fc042598ca016d1b2b60707db131e53ef4c504