Rename starlark(type) to starlark(dot_type)
Summary: In Starlark modules we have a parameter attribute `starlark(type)` which provides a docs-only type. We then have a function-level `starlark(type)` which provides the `foo.type` attribute. Having the same name for these was a bad plan. We should rename the `foo.type` one. I thought about `type_attribute` but there is already an `attribute` annotation on a function that it gets confusing with. Using `dot_type` is unambiguous.

Reviewed By: stepancheg

Differential Revision: D45601538

fbshipit-source-id: aff599999e3af1f6a9d3b9c4660f003baea0cf33